requirements.txt
einops
einops-exts
bitsandbytes
accelerate
transformers
peft
bigmodelvis
torch
torchvision
pillow
more-itertools
datasets
braceexpand
webdataset
wandb
nltk
scipy
inflection
sentencepiece
open_clip_torch
mmengine
gradio
evaluate==0.4.0
scikit-learn==1.2.2
googletrans==3.1.0a0